Update builders test
