+++ package.json
@@ -27,6 +27,7 @@
   "scripts": {
     "build": "unbuild",
     "dev": "vitest dev",
+    "dev:ui": "vitest dev --ui",
     "lint": "eslint --cache --ext .ts,.js,.mjs,.cjs . && prettier -c src test",
     "lint:fix": "eslint --cache --ext .ts,.js,.mjs,.cjs . --fix && prettier -c src test -w",
     "prepack": "pnpm run build",
@@ -42,6 +43,7 @@
   "devDependencies": {
     "@types/node": "^18.15.11",
     "@vitest/coverage-c8": "^0.29.8",
+    "@vitest/ui": "^0.29.8",
     "changelogen": "^0.5.2",
     "eslint": "^8.37.0",
     "eslint-config-unjs": "^0.1.0",
